refactor: update token transfer structure
