+++ token-transfer/tests/machines.test.ts
@@ -2,8 +2,8 @@ import { StateMachine } from "@stackr/sdk/machine";
 import { expect } from "chai";
 import { Wallet, ZeroHash, verifyTypedData } from "ethers";
 import genesisState from "../genesis-state.json";
-import { schemas } from "../src/actions";
-import { ERC20, Leaves } from "../src/state";
+import { schemas } from "../src/stackr/actions";
+import { ERC20, Leaves } from "../src/stackr/state";
 import { transitions } from "../src/transitions";
 import { stackrConfig } from "../stackr.config";
